# [16.0] Various easy performance boosts