@hokolomopo 2023-04-13
Fixed a bug in `splitNumber()`, I didn't handle correctly the case where we round up and the last digit is a "9". Added a test. Unfortunately, it makes the function more complex :/ I'd really like to use simply Number.toFixed(), but it has some strange illogical behavour...

TODO for tomorrow: also check if it's worth changing `range()` to an iterator, to avoid allocating arrays where we don't have to.